Aggregating common scheduler components into a parent class for Schedulers
